Add tag implementation into article Add scroll to tabs on discover page
