+++ src/components/views/article2.js
@@ -1,6 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import moment from "moment";
+import { Link } from "react-router-dom";
 
 import { Item } from "semantic-ui-react";
 
@@ -14,11 +15,14 @@ const Article = (props) => {
   return (
     <Item className="article--2">
       <Item.Content>
-        <Item.Header as="a">{article.title}</Item.Header>
+        <Item.Header as={Link} to={`/article/${article.slug}`}>{article.title}</Item.Header>
         <Item.Meta>
           <span className="right floated date">{`${createdTime} ago`}</span>
           <span className="ui black">{`${article.time_to_read} min read`}</span>
         </Item.Meta>
+        <Item.Description>
+          Created by <a as={Link} to="/p">{article.author.username}</a>
+        </Item.Description>
         <Item.Description>
           {article.description}
         </Item.Description>
